--- src/data/initialbids.ts
@@ -1,114 +1,0 @@
-import { BigNumber } from 'ethers'
-
-export const initialBid = [
-  {
-    address: '0xf1',
-    tokenOut: BigNumber.from(1000), // price: 1.3
-    tokenIn: BigNumber.from(769.230769230769),
-  },
-  {
-    address: '0xf2',
-    tokenOut: BigNumber.from(15000), // price: 1.2
-    tokenIn: BigNumber.from(12500),
-  },
-  {
-    address: '0xf3',
-    tokenOut: BigNumber.from(1000), // price: 1.2
-    tokenIn: BigNumber.from(833.333333333333),
-  },
-  {
-    address: '0xf4',
-    tokenOut: BigNumber.from(1000), // price: 1.15
-    tokenIn: BigNumber.from(869.565217391304),
-  },
-  {
-    address: '0xf5',
-    tokenOut: BigNumber.from(1000), // price: 1.1
-    tokenIn: BigNumber.from(909.090909090909),
-  },
-  {
-    address: '0xf6',
-    tokenOut: BigNumber.from(10000), // price: 1.05
-    tokenIn: BigNumber.from(9523.80952380952),
-  },
-  {
-    address: '0xf7',
-    tokenOut: BigNumber.from(1000), // price: 1.05
-    tokenIn: BigNumber.from(952.380952380952),
-  },
-  {
-    address: '0xf8',
-    tokenOut: BigNumber.from(5000), // price: 0.95
-    tokenIn: BigNumber.from(5263.15789473684),
-  },
-  {
-    address: '0xf9',
-    tokenOut: BigNumber.from(500), // price: 0.91
-    tokenIn: BigNumber.from(549.450549450549),
-  },
-  {
-    address: '0xf10',
-    tokenOut: BigNumber.from(10000), // price: 0.85
-    tokenIn: BigNumber.from(11764.7058823529),
-  },
-  {
-    address: '0xf11',
-    tokenOut: BigNumber.from(2000), // price: 0.77
-    tokenIn: BigNumber.from(2597.4025974026),
-  },
-  {
-    address: '0xf12',
-    tokenOut: BigNumber.from(2000), // price: 0.765
-    tokenIn: BigNumber.from(2614.37908496732),
-  },
-  {
-    address: '0xf13',
-    tokenOut: BigNumber.from(2000), // price: 0.76
-    tokenIn: BigNumber.from(2631.57894736842),
-  },
-  {
-    address: '0xf14',
-    tokenOut: BigNumber.from(2000), // price: 0.755
-    tokenIn: BigNumber.from(2649.00662251656),
-  },
-  {
-    address: '0xf15',
-    tokenOut: BigNumber.from(10000), // price: 0.75
-    tokenIn: BigNumber.from(13333.3333333333),
-  },
-  {
-    address: '0xf16',
-    tokenOut: BigNumber.from(2000), // price: 0.75
-    tokenIn: BigNumber.from(2666.66666666667),
-  },
-  {
-    address: '0xf17',
-    tokenOut: BigNumber.from(1000), // price: 0.7
-    tokenIn: BigNumber.from(1428.57142857143),
-  },
-  {
-    address: '0xf18',
-    tokenOut: BigNumber.from(5000), // price: 0.65
-    tokenIn: BigNumber.from(7692.30769230769),
-  },
-  {
-    address: '0xf19',
-    tokenOut: BigNumber.from(1000), // price: 0.6
-    tokenIn: BigNumber.from(1666.66666666667),
-  },
-  {
-    address: '0xf20',
-    tokenOut: BigNumber.from(1000), // price: 0.5
-    tokenIn: BigNumber.from(2000),
-  },
-  {
-    address: '0xf21',
-    tokenOut: BigNumber.from(1000), // price: 0.5
-    tokenIn: BigNumber.from(2000),
-  },
-  {
-    address: '0xf22',
-    tokenOut: BigNumber.from(1000), // price: 0.2
-    tokenIn: BigNumber.from(5000),
-  },
-]
